refactor(hir): scope
