+++ crates/hir/src/hir_def/module.rs
@@ -265,11 +265,9 @@ impl LowerModuleCtx<'_> {
 
     fn lower_class_decl(&mut self, class_decl: ast::ClassDeclaration) -> ClassId {
         let container_id = ContainerId::ModuleId(self.module_id);
-        let class_def = lower_class_def(
-            class_decl.clone(),
-            container_id,
-            |ty| self.expr_ctx().lower_data_ty(ty),
-        );
+        let class_def = lower_class_def(class_decl.clone(), container_id, |ty| {
+            self.expr_ctx().lower_data_ty(ty)
+        });
 
         alloc_idx_and_src! {
             class_def => self.module.classes,
